Move `isGPUResourceKey` to `utils/resources.go` for reusability
