+++ ray-operator/controllers/ray/common/pod.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"regexp"
 	"sort"
 	"strconv"
 	"strings"
@@ -854,7 +853,7 @@ func addWellKnownAcceleratorResources(rayStartParams map[string]string, resource
 
 		// Scan for resource keys of gpus
 		if _, ok := rayStartParams["num-gpus"]; !ok {
-			if isGPUResourceKey(resourceKeyString) && !resourceValue.IsZero() {
+			if utils.IsGPUResourceKey(resourceKeyString) && !resourceValue.IsZero() {
 				rayStartParams["num-gpus"] = strconv.FormatInt(resourceValue.Value(), 10)
 			}
 		}
@@ -1026,15 +1025,4 @@ func findMemoryReqOrLimit(container corev1.Container) (res *resource.Quantity) {
 		return mem
 	}
 	return nil
-}
-
-func isGPUResourceKey(key string) bool {
-	// ending with "gpu" like "nvidia.com/gpu"
-	if strings.HasSuffix(key, "gpu") {
-		return true
-	}
-	// Nvidia Multi-Instance GPU in the form of "nvidia.com/mig-<slice_count>g.<memory_size>gb" like "nvidia.com/mig-2g.32gb"
-	// reference: https://github.com/NVIDIA/k8s-device-plugin#configuration-option-details
-	match, _ := regexp.MatchString(`nvidia\.com/mig-\d+g\.\d+gb$`, key)
-	return match
 }
